feat(dht): separate ClientData from DhtNode
